pyproject.toml
[project]
name = "sb1"
version = "0.1.0"
description = ""
authors = [
    {name = "Alexei",email = "streiker.saik@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "python-dotenv (>=1.1.1,<2.0.0)",
    "psycopg2 (>=2.9.10,<3.0.0)",
    "django (>=5.2.6,<6.0.0)",
    "djangorestframework (>=3.16.1,<4.0.0)",
    "djangorestframework-simplejwt (>=5.5.1,<6.0.0)",
    "django-filter (>=25.1,<26.0)",
    "drf-yasg (>=1.21.10,<2.0.0)",
    "django-cors-headers (>=4.9.0,<5.0.0)",
    "pillow (>=11.3.0,<12.0.0)",
    "celery (>=5.5.3,<6.0.0)",
    "eventlet (>=0.40.3,<0.41.0)",
    "redis (>=6.4.0,<7.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.lint.dependencies]
flake8 = "^7.3.0"
isort = "^6.0.1"
black = "^25.9.0"
mypy = "^1.18.2"


[tool.poetry.group.dev.dependencies]
pytest = "^8.4.2"
pytest-django = "^4.11.1"
pytest-cov = "^7.0.0"

[tool.black]
# Максимальная длина строки
line-length = 119
# Файлы, которые не нужно форматировать
exclude = '''
(
  /(
      \.git
  )/
)
'''

[tool.isort]
# максимальная длина строки
line_length = 119

[tool.mypy]
disallow_untyped_defs = true
warn_return_any = true
exclude = 'venv'